Move kolena.initialize unit tests to appropriate file
